App Center: correctly use repo's latest version for baseUrl
